+++ CHANGELOG.md
@@ -34,6 +34,8 @@
 - Update `get_corrected_centroid_guesses()` to also extract from base clusters.
 - Remove extraneous return values from `PolisClusteringResult`.
 - Add `data_presenter.generate_figure_polis()` for making graphs from PolisClusteringResult.
+- Add `group_aware_consensus` dataframe to PolisClusteringResult of polis implementation.
+
 
 ### Chores
 - Moved agora implementation from `reddwarf.agora` to `reddwarf.implementations.agora` (deprecation warning).